+++ DeliveryService/src/main/java/anna/deliveryservice/repository/OrderRepository.java
@@ -1,18 +1,13 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
 package anna.deliveryservice.repository;
 
 import anna.deliveryservice.domain.Order;
 
 /**
- *
- * @author Alex
+ * @author Anna
+ * Interface for repository classes of order entity
  */
 public interface OrderRepository {
     Order save(Order order);
-    Order findById(int id);
+    Order findById(Long id);
     Order update(Order order);
 }
