+++ DeliveryService/src/main/java/anna/deliveryservice/service/PizzaService.java
@@ -1,16 +1,12 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
 package anna.deliveryservice.service;
 
 import anna.deliveryservice.domain.Pizza;
 
 /**
- *
- * @author Alex
+ * @author Anna
+ * Interface for service classes working with pizza
  */
 public interface PizzaService {
-    Pizza find(Integer id);
+    
+    Pizza find(Long id);
 }
